+++ dash_app/callbacks/consensustable.py
@@ -1,7 +1,7 @@
 from dash.dependencies import Input, Output, State
 
 from dash_app.components import consensustable, consensustree
-from ..components import jsontools
+from ..components import tools
 from ..layout.layout_ids import *
 
 from ..server import app
@@ -15,14 +15,14 @@ from ..server import app
 def update_partial_table_data(jsonified_full_consensustable: str, jsonified_tree: str, slider_value: float):
     if not jsonified_full_consensustable or not jsonified_tree:
         return []
-    full_consensustable_data = jsontools.unjsonify_df(jsonified_full_consensustable)
-    full_consensustree_data = jsontools.unjsonify_builtin_types(jsonified_tree)
+    full_consensustable_data = tools.unjsonify_df(jsonified_full_consensustable)
+    full_consensustree_data = tools.unjsonify_builtin_types(jsonified_tree)
     full_consensustree_tree = consensustree.dict_to_tree(full_consensustree_data)
     table_without_consensuses_smaller_than_slider = consensustable.remove_smaller_than_slider(full_consensustable_data,
                                                                                               full_consensustree_tree,
                                                                                               slider_value)
     # tu dołożyć ewentualne zmiany wynikające z innych elementów Input/State
-    return jsontools.jsonify_df(table_without_consensuses_smaller_than_slider)
+    return tools.jsonify_df(table_without_consensuses_smaller_than_slider)
 
 @app.callback(
     Output(id_consensuses_table, 'data'),
@@ -31,7 +31,7 @@ def update_partial_table_data(jsonified_full_consensustable: str, jsonified_tree
 def to_consensustable_content(jsonified_partial_consensustable):
     if not jsonified_partial_consensustable:
         return []
-    partial_consensustable_data = jsontools.unjsonify_df(jsonified_partial_consensustable)
+    partial_consensustable_data = tools.unjsonify_df(jsonified_partial_consensustable)
     data_rows = partial_consensustable_data.to_dict("rows")
     return data_rows
 
@@ -42,7 +42,7 @@ def to_consensustable_content(jsonified_partial_consensustable):
 def update_columns(jsonified_partial_consensustable):
     if not jsonified_partial_consensustable:
         return [{}]
-    partial_consensustable_data = jsontools.unjsonify_df(jsonified_partial_consensustable)
+    partial_consensustable_data = tools.unjsonify_df(jsonified_partial_consensustable)
     return [{"name": i, "id": i} for i in partial_consensustable_data.columns]
 
 @app.callback(
@@ -53,8 +53,8 @@ def update_columns(jsonified_partial_consensustable):
 def color_consensuses_table_cells(jsonified_partial_consensustable, jsonified_consensus_tree):
     if not jsonified_partial_consensustable or not jsonified_consensus_tree:
         return []
-    partial_consensustable_data = jsontools.unjsonify_df(jsonified_partial_consensustable)
-    consensustree_data = jsontools.unjsonify_builtin_types(jsonified_consensus_tree)
+    partial_consensustable_data = tools.unjsonify_df(jsonified_partial_consensustable)
+    consensustree_data = tools.unjsonify_builtin_types(jsonified_consensus_tree)
     tree = consensustree.dict_to_tree(consensustree_data)
 
     return consensustable.get_cells_styling(tree, partial_consensustable_data)
