+++ dash_app/callbacks/consensustree.py
@@ -1,7 +1,7 @@
 from dash.dependencies import Input, Output
 
 from dash_app.components import consensustable
-from ..components import jsontools
+from ..components import tools
 from ..layout.layout_ids import *
 from ..components import consensustree
 from ..server import app
@@ -13,10 +13,10 @@ from ..server import app
 def update_current_tree_state(jsonified_full_consensustree):
     if not jsonified_full_consensustree:
         return []
-    full_consensustree_data = jsontools.unjsonify_builtin_types(jsonified_full_consensustree)
+    full_consensustree_data = tools.unjsonify_builtin_types(jsonified_full_consensustree)
     full_consensus_tree = consensustree.dict_to_tree(full_consensustree_data)
     current_consensustree_data = consensustree.tree_to_dict(full_consensus_tree)
-    return jsontools.jsonify_builtin_types(current_consensustree_data)
+    return tools.jsonify_builtin_types(current_consensustree_data)
 
 
 @app.callback(
@@ -28,9 +28,9 @@ def update_current_tree_state(jsonified_full_consensustree):
 def to_consensustree_graph(jsonified_current_consensustree, slider_value, leaf_info, jsonified_full_consensustable):
     if not jsonified_current_consensustree or not jsonified_full_consensustable:
         return {}
-    current_consensustree_data = jsontools.unjsonify_builtin_types(jsonified_current_consensustree)
+    current_consensustree_data = tools.unjsonify_builtin_types(jsonified_current_consensustree)
     current_consensustree_tree = consensustree.dict_to_tree(current_consensustree_data)
-    full_consensustable_data = jsontools.unjsonify_df(jsonified_full_consensustable)
+    full_consensustable_data = tools.unjsonify_df(jsonified_full_consensustable)
     graph = consensustree.get_consensustree_graph(current_consensustree_tree, slider_value, leaf_info, full_consensustable_data)
     return graph
 
@@ -56,11 +56,11 @@ def to_consensus_node_details_table(tree_click_data, jsonified_full_consensustab
         return []
     clicked_node = tree_click_data['points'][0]
     node_id = clicked_node['pointIndex']
-    full_consensustable = jsontools.unjsonify_df(jsonified_full_consensustable)
-    consensustree_data = jsontools.unjsonify_builtin_types(jsonified_consensustree)
+    full_consensustable = tools.unjsonify_df(jsonified_full_consensustable)
+    consensustree_data = tools.unjsonify_builtin_types(jsonified_consensustree)
     tree = consensustree.dict_to_tree(consensustree_data)
     node_details_df = consensustable.get_consensus_details_df(node_id, full_consensustable, tree)
-    return jsontools.jsonify_df(node_details_df)
+    return tools.jsonify_df(node_details_df)
 
 @app.callback(
     Output(id_consensus_node_details_table, 'data'),
@@ -69,7 +69,7 @@ def to_consensus_node_details_table(tree_click_data, jsonified_full_consensustab
 def to_consensusnode_details_content(jsonified_consensus_details_table):
     if not jsonified_consensus_details_table:
         return []
-    consensus_details_table_data = jsontools.unjsonify_df(jsonified_consensus_details_table)
+    consensus_details_table_data = tools.unjsonify_df(jsonified_consensus_details_table)
     data_rows = consensus_details_table_data.to_dict("rows")
     return data_rows
 
@@ -83,7 +83,7 @@ def to_consensus_node_details_distribution(tree_click_data, jsonified_full_conse
         return ""
     clicked_node = tree_click_data['points'][0]
     node_id = clicked_node['pointIndex']
-    full_consensustable = jsontools.unjsonify_df(jsonified_full_consensustable)
+    full_consensustable = tools.unjsonify_df(jsonified_full_consensustable)
     distribution_figure = consensustable.get_node_distribution_fig(node_id, full_consensustable)
     return distribution_figure
 
@@ -94,7 +94,7 @@ def to_consensus_node_details_distribution(tree_click_data, jsonified_full_conse
 def update_columns(jsonified_consensus_details_table):
     if not jsonified_consensus_details_table:
         return [{}]
-    consensus_details_table_data = jsontools.unjsonify_df(jsonified_consensus_details_table)
+    consensus_details_table_data = tools.unjsonify_df(jsonified_consensus_details_table)
     return [{"name": i, "id": i} for i in list(consensus_details_table_data.columns)]
 
 @app.callback(
@@ -123,7 +123,7 @@ def show_consensus_tree_info(click_data):
 def to_consensustree_leaf_info_options_dropdown(jsonified_full_consensustable):
     if not jsonified_full_consensustable:
         return []
-    full_consensustable = jsontools.unjsonify_df(jsonified_full_consensustable)
+    full_consensustable = tools.unjsonify_df(jsonified_full_consensustable)
     metadata = consensustable.get_metadata_list(full_consensustable)
     dropdown_options = consensustree.get_leaf_info_dropdown_options(metadata)
     return dropdown_options
