+++ dash_app/callbacks/parameters.py
@@ -1,6 +1,6 @@
 from dash.dependencies import Input, Output
 import dash_html_components as html
-from ..components import jsontools
+from ..components import tools
 from ..layout.layout_ids import *
 
 from ..server import app
@@ -12,7 +12,7 @@ from ..server import app
 def update_pangenome_parameters(jsonified_params):
     if not jsonified_params:
         return []
-    parameters = jsontools.unjsonify_builtin_types(jsonified_params)
+    parameters = tools.unjsonify_builtin_types(jsonified_params)
     params_info = []
     for param_name, param_value in parameters.items():
         params_info.append(html.P(f"{param_name}: {param_value}"))
